requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
motor==3.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0

# AI Model dependencies
tensorflow==2.16.1
Pillow==10.3.0
groq==0.5.0
numpy==1.26.4
typing-extensions==4.11.0
